+++ intel-mkl-tool/Cargo.toml
@@ -11,20 +11,23 @@ license     = "MIT"
 
 [features]
 default = ["cli"]
-cli = ["structopt"]
+archive = ["curl", "tar", "zstd"]
+cli     = ["structopt", "archive"]
 
 [dependencies]
 anyhow = "1.0.31"
-curl = "0.4.29"
 derive_more = "0.99.8"
 dirs = "2.0.2"
 glob = "0.3.0"
 pkg-config = "0.3.17"
-tar = "0.4.29"
-zstd = "0.5.3"
+
+# archive
+curl = { version = "0.4.29", optional = true }
+tar  = { version = "0.4.29", optional = true }
+zstd = { version = "0.5.3",  optional = true }
 
 # CLI
-structopt  = { version = "0.3.15", optional = true }
+structopt = { version = "0.3.15", optional = true }
 
 [dev-dependencies]
 paste = "0.1.17"
